+++ yew-ui/src/components/host.rs
@@ -22,7 +22,7 @@ use futures::channel::mpsc;
 use gloo_timers::callback::Timeout;
 use log::debug;
 use videocall_client::{create_microphone_encoder, MicrophoneEncoderTrait};
-use videocall_client::{CameraEncoder, MediaDeviceList, ScreenEncoder};
+use videocall_client::{CameraEncoder, MediaDeviceList, ScreenEncoder, ScreenShareEvent};
 use videocall_types::protos::media_packet::media_packet::MediaType;
 use yew::prelude::*;
 
@@ -119,6 +119,10 @@ pub struct MeetingProps {
     /// The parent should disable the mic and optionally display an error.
     #[prop_or_default]
     pub on_microphone_error: Callback<String>,
+
+    /// Called when screen share state changes (started, cancelled, stopped).
+    /// This allows the parent component to react to screen share lifecycle events.
+    pub on_screen_share_state: Callback<ScreenShareEvent>,
 }
 
 impl Component for Host {
@@ -156,7 +160,13 @@ impl Component for Host {
         );
 
         let screen_bitrate = screen_bitrate_kbps().unwrap_or(1000);
-        let mut screen = ScreenEncoder::new(client.clone(), screen_bitrate, screen_callback);
+        let screen_state_callback = ctx.props().on_screen_share_state.clone();
+        let mut screen = ScreenEncoder::new(
+            client.clone(),
+            screen_bitrate,
+            screen_callback,
+            screen_state_callback,
+        );
 
         let (tx, rx) = mpsc::unbounded();
         client.subscribe_diagnostics(tx.clone(), MediaType::VIDEO);
@@ -243,7 +253,8 @@ impl Component for Host {
         // TODO: use atomic bools for the encoders
         self.client.set_audio_enabled(self.mic_enabled);
         self.client.set_video_enabled(self.video_enabled);
-        self.client.set_screen_enabled(self.share_screen);
+        // Note: set_screen_enabled is now handled internally by ScreenEncoder
+        // based on actual stream state (Started/Stopped callbacks)
 
         if first_render {
             ctx.link().send_message(Msg::Start);